Fixing a bunch of bugs while trying to run for the first time.
