+++ bin/dynamo-consistency
@@ -1,9 +1,11 @@
 #! /usr/bin/env python
 
 
-import logging
+import os
 import time
+import logging
 import datetime
+import traceback
 
 
 from dynamo_consistency import opts
@@ -21,7 +23,7 @@ if __name__ == '__main__':
     config.LOCATION = opts.CONFIG
 
     config_dict = config.config_dict()
-    logdir = config['LogLocation']
+    logdir = config_dict['LogLocation']
 
     global_log = os.path.join(logdir, 'consistency.log')
     logsetup.change_logfile(global_log)
@@ -30,8 +32,12 @@ if __name__ == '__main__':
     logger.info('Starting run on %s', site)
 
     try:
+        weblog = os.path.join(config_dict['WebDir'], '%s.log' % site)
+        if os.path.exists(weblog):
+            os.remove(weblog)
+
         logsetup.change_logfile(
-            os.path.join(config_dict['WebDir'], '%s.log' % site),
+            os.path.join(weblog),
             os.path.join(logdir, site, '%s.log' % 
                          datetime.datetime.fromtimestamp(time.time()).strftime('%y%m%d_%H%M%S'))
             )
@@ -40,7 +46,10 @@ if __name__ == '__main__':
 
     except Exception as e:
 
-        LOG.error(e)
+        full = traceback.format_exc()
+
+        logger.error(full)
+        logger.error(e)
 
         if opts.EMAIL:
             from cmstoolbox.emailtools import send_mail
@@ -50,7 +59,7 @@ if __name__ == '__main__':
                 sender=mail_config['Sender'],
                 recipients=mail_config['Recipients'],
                 subject='Exception in dynamo-consistency',
-                message_text=e)
+                message_text=full)
 
     summary.unlock_site(site)
