+++ dynamo_consistency/dynamo/siteinfo.py
@@ -42,7 +42,7 @@ def ready_sites():
     global _READY    #pylint: disable=global-statement
 
     if _READY is None:
-        _READY = set(_small_query('SELECT name FROM sites WHERE status = `ready`'))
+        _READY = set(_small_query('SELECT name FROM sites WHERE status = "ready"'))
 
     return _READY
